+++ prototype/nplinker/class_info/chem_classes.py
@@ -14,7 +14,10 @@
 
 import os
 import glob
+from collections import Counter
+
 from canopus.classifications_to_gnps import analyse_canopus
+from canopus import Canopus
 
 from ..logconfig import LogConfig
 
@@ -38,7 +41,6 @@ class ChemClassPredictions:
             mne_dir: str, mne_dir found in root_dir of nplinker project
             gnps_dir: str, root dir where all gnps info is found
         """
-        # todo: use canopus_treemap to convert canopus result
         self._canopus = CanopusResults(canopus_dir, gnps_dir)
         self._molnetenhancer = MolNetEnhancerResults(mne_dir)
 
@@ -87,11 +89,25 @@ class CanopusResults:
 
         If possible, convert canopus output with canopus_treemap. Otherwise
         canopus output is used directly. If converting fails, probably the MN
-        version is too old.
+        version is too old, and canopus output is read directly
         """
+        self._canopus_dir = canopus_dir
+        self._gnps_dir = gnps_dir
+        self._molfam_classes, self._molfam_classes_names, \
+            self._molfam_classes_names_inds = (None, None, None)
+        self._spectra_classes, self._spectra_classes_names, \
+            self._spectra_classes_names_inds = (None, None, None)
+        if os.path.isdir(self._canopus_dir):
+            self._read_all_classes()
+        else:
+            logger.info(f"No CANOPUS results present at {self._canopus_dir}. "
+                        f"(set run_canopus=true in the .toml to run CANOPUS)")
+
+    def _read_all_classes(self):
+        """Wrapper to read all canopus output and store them in the object"""
         ci_file = False  # find cluster_index_classifications in canopus_dir
         compi_file = False  # same for component_index_classifications
-        canopus_files = glob.glob(os.path.join(canopus_dir, "*"))
+        canopus_files = glob.glob(os.path.join(self._canopus_dir, "*"))
         possible_ci_file = [canopus_file for canopus_file in canopus_files
                             if canopus_file.endswith(
                                 'cluster_index_classifications.txt')]
@@ -106,13 +122,19 @@ class CanopusResults:
         if not os.path.isfile(ci_file):
             logger.info('Converting canopus output using canopus_treemap')
             try:
-                analyse_canopus(canopus_dir, gnps_dir, canopus_dir)
-            except Exception as e:
-                logger.warning(f'canopus_treemap failed with: {e}. '
-                               f'Probably the MN version from GNPS is too old')
+                analyse_canopus(self._canopus_dir, self._gnps_dir,
+                                self._canopus_dir)
+            except FileNotFoundError as er1:
+                logger.warning(
+                    f'{er1}. CANOPUS output is missing in {self._canopus_dir}')
+            except Exception as er2:
+                logger.warning(
+                    f'canopus_treemap failed with: {er2}. Probably the MN '
+                    f'version from GNPS is too old. Will attempt to read '
+                    f'directly from canopus_dir')
             else:
                 # find processed canopus files again
-                canopus_files = glob.glob(os.path.join(canopus_dir, "*"))
+                canopus_files = glob.glob(os.path.join(self._canopus_dir, "*"))
                 possible_ci_file = [
                     canopus_file for canopus_file in canopus_files if
                     canopus_file.endswith('cluster_index_classifications.txt')]
@@ -125,25 +147,25 @@ class CanopusResults:
                 if possible_compi_file:
                     compi_file = possible_compi_file[0]
 
-        # todo: directly use canopus output when canopus_treemap fails
         if os.path.isfile(ci_file):
             spectra_classes_names, spectra_classes = \
                 self._read_spectra_classes(ci_file)
 
-            molfam_classes_names, molfam_classes = self._read_molfam_classes(
-                compi_file)
-            self._molfam_classes = molfam_classes
-            self._molfam_classes_names = molfam_classes_names
-            self._molfam_classes_names_inds = {elem: i for i, elem in
-                                               enumerate(molfam_classes_names)}
-
+            if os.path.isfile(compi_file):
+                molfam_classes_names, molfam_classes = self.\
+                    _read_molfam_classes(compi_file)
+                self._molfam_classes = molfam_classes
+                self._molfam_classes_names = molfam_classes_names
+                self._molfam_classes_names_inds = {
+                    elem: i for i, elem in enumerate(molfam_classes_names)}
         else:
             # use canopus output correctly (only for spectra)
             logger.info("Attempting to read spectra classes directly from "
                         "canopus_dir (canopus_summary.tsv)")
             spectra_classes_names, spectra_classes = \
-                self._read_spectra_classes_directly(canopus_dir)
-            # todo: add function so that classes can be determined for molfams
+                self._read_spectra_classes_directly()
+            # molfams have to be added later with info about molfam <- spectra
+            # this happens with transfer_spec_classes_to_molfams() in loader.py
 
         self._spectra_classes = spectra_classes
         self._spectra_classes_names = spectra_classes_names
@@ -158,9 +180,9 @@ class CanopusResults:
         Returns:
             Tuple of:
             - ci_classes_names: list of str - the names of each different level
-            - ci_classes: dict of {str, lists of tuple(str, float)} - per spectrum (key) the classes for each level
+            - ci_classes: dict of {str, lists of tuple(str, float)} - per spectrum index (key) the classes for each level
                 where each level is a list of (class_name, score) sorted on best choice so index 0 is the best
-                class prediction for a level
+                class prediction for a level. When no class is present, instead of Tuple it will be None for that level.
         """
         ci_classes = {}  # make a dict {ci: [[(class,score)]]}
         ci_classes_header = None
@@ -197,47 +219,106 @@ class CanopusResults:
                                 ci_classes_header[-3:]]
         return ci_classes_names, ci_classes
 
-    def _read_spectra_classes_directly(self, canopus_dir):
-        """Read canopus classes directly from canopus output
+    def _read_spectra_classes_directly(self):
+        """Read canopus classes directly from canopus output and save output
 
-        Args:
-            canopus_dir: str, canopus dir
         Returns:
             Tuple of:
             - can_classes_names: list of str - the names of each different level
-            - can_classes: dict of {str, lists of tuple(str, float)} - per spectrum (key) the classes for each level
+            - can_classes: dict of {str, lists of tuple(str, float/None)} - per spectrum index (key) the classes for each level
                 where each level is a list of (class_name, score) sorted on best choice so index 0 is the best
-                class prediction for a level
+                class prediction for a level. When no class is present, instead of Tuple it will be None for that level.
         """
-        can_classes = {}  # make a dict {can: [[(class,score)]]}
+        can_classes = {}  # make a dict {spec: [[(class,score)]]}
         can_classes_header = None
         can_classes_names = []
 
-        input_file = os.path.join(canopus_dir, 'canopus_summary.tsv')
+        input_file = os.path.join(self._canopus_dir, 'canopus_summary.tsv')
+        sep = '\t'
         if os.path.isfile(input_file):
             logger.info(f"reading canopus results for spectra directly from "
                         f"{input_file}")
             with open(input_file) as inf:
-                can_classes_header = inf.readline().strip().split("\t")
+                can_classes_header = inf.readline().strip().split(sep)
                 # get [4:8] - level 5 subclass        class   superclass
                 starti = 4
                 stopi = 8
                 for line in inf:
-                    line = line.strip('\n').split("\t")
-                    # make tuples with scores as None
-                    classes_list = [(cls, None) for cls in line[starti:stopi]]
+                    line = line.strip('\n').split(sep)
+                    # make tuples with scores as 1
+                    classes_list = [[(cls, 1)] if cls else [None]
+                                    for cls in line[starti:stopi]]
                     # assuming that all MGFs have ScanNumber headers..
                     spec_index = line[0].split('ScanNumber')[-1]
                     classes_list.reverse()  # reverse so superclass is first
                     can_classes[spec_index] = classes_list
+
+            # use canopus_treemap to produce NPClassifier classes
+            # TODO: probably change when sirius v5 comes out
+            logger.debug("Using canopus_treemap to get NPC classes")
+            canopus_workspace = Canopus(sirius=self._canopus_dir)
+            npc_file = os.path.join(self._canopus_dir, "npc_summary.tsv")
+            canopus_workspace.npcSummary().to_csv(npc_file, sep=sep)
+            with open(npc_file) as inf:
+                npc_can_classes_header = inf.readline().strip().split(sep)
+                # get [2:7] - name,directoryName,pathway,pathwayProbability,superclass,superclassProbability,class,classProbability,ClassyFirePrediction,ClassyFirePredictionProbability
+                npc_starti = 2
+                npc_stopi = 7
+                for line in inf:
+                    line = line.strip('\n').split(sep)
+                    # make tuples with scores
+                    npc_classes_list = [
+                        [(cls, float(c_score))] if cls != 'N/A' else [None]
+                        for cls, c_score in
+                        zip(line[npc_starti:npc_stopi:2],
+                            line[npc_starti+1:npc_stopi+1:2])]
+                    # assuming that all MGFs have ScanNumber headers..
+                    spec_index = line[1].split('ScanNumber')[-1]
+                    can_classes[spec_index].extend(npc_classes_list)
         else:
             logger.warn(f'could not load canopus results for spectra; '
                         f'{input_file} missing from canopus_dir')
 
         if can_classes_header:
+            # important that the names here match with MIBiG
             can_classes_names = [f"cf_{elem}" for elem in
                                  can_classes_header[starti:stopi]]
             can_classes_names.reverse()  # reverse so superclass is first
+            npc_can_classes_names = [
+                f"npc_{elem}" for elem in
+                npc_can_classes_header[npc_starti:npc_stopi:2]]
+            can_classes_names.extend(npc_can_classes_names)
+
+            # make sure that non-existing NPC predictions are added as [None]
+            # to all spectra - some spectra where not included in npc_summary
+            len_classes_names = len(can_classes_names)
+            for spec_index, classes_list in can_classes.items():
+                diff = len_classes_names - len(classes_list)
+                if diff > 0:
+                    for _ in range(diff):
+                        classes_list.append([None])
+
+            # save output, next time clusterindex_classifications.txt is read
+            # save just the info that we need in the format that is parsed in
+            # read_spectra_classes()
+            output_file = os.path.join(
+                self._canopus_dir, "cluster_index_classifications.txt")
+            with open(output_file, 'w') as outf:
+                header = ['componentindex', 'cluster index', 'formula'] +\
+                    [name.partition('_')[-1] for name in can_classes_names]
+                outf.write('\t'.join(header)+'\n')
+                for spec, classes in can_classes.items():
+                    formatted_classes = []
+                    for lvl in classes:
+                        lvl_classes = []
+                        for tup in lvl:
+                            if not tup:
+                                lvl_classes.append('')
+                            else:
+                                lvl_classes.append(f"{tup[0]}:{tup[1]:.3f}")
+                        formatted_classes.append('; '.join(lvl_classes))
+                    output_l = ['-', spec, '-'] + formatted_classes
+                    outf.write('\t'.join(output_l)+'\n')
         return can_classes_names, can_classes
 
     def _read_molfam_classes(self, input_file):
@@ -248,9 +329,9 @@ class CanopusResults:
         Returns:
             Tuple of:
             - compi_classes_names: list of str - the names of each different level
-            - compi_classes: dict of {str: lists of tuple(str, float)} - per molfam (key) the classes for each level
+            - compi_classes: dict of {str: lists of tuple(str, float)} - per molfam index (key) the classes for each level
                 where each level is a list of (class_name, fraction) sorted on best choice so index 0 is the best
-                class prediction for a level
+                class prediction for a level. When no class is present, instead of Tuple it will be None for that level.
         """
         compi_classes = {}  # make a dict {compi: [[(class,score)]]}
         compi_classes_header = None
@@ -284,6 +365,84 @@ class CanopusResults:
                                   [f"npc_{elem}" for elem in
                                    compi_classes_header[-3:]]
         return compi_classes_names, compi_classes
+
+    def transfer_spec_classes_to_molfams(self, molfams, fraction_cutoff=0.):
+        """Set _molfam_classes(_names) from spectra_classes and return classes
+
+        This can be used in the _loader to get molfam classes when the GNPS MN
+        version is too old and canopus_treemap fails to work directly.
+
+        Args:
+            molfams: list of MolecularFamily from the NPLinker space
+            fraction_cutoff: float, cut-off for the fraction of class terms
+                needed to be included in the molfam
+        Returns:
+            dict of {str: lists of tuple(str, float)} - per molfam (key) the classes for each level
+                where each level is a list of (class_name, fraction) sorted on best choice so index 0 is the best
+                class prediction for a level. When no class is present, instead of Tuple it will be None for that level.
+        """
+        self._molfam_classes_names = self._spectra_classes_names
+        self._molfam_classes_names_inds = self._spectra_classes_names_inds
+        molfam_classes = {}
+
+        for molfam in molfams:
+            fid = str(molfam.family_id)  # the key
+            spectra = molfam.spectra
+            # if singleton family, format like '-1_spectrum-id'
+            if fid == '-1':
+                spec_id = spectra[0].spectrum_id
+                fid += f'_{spec_id}'
+            len_molfam = len(spectra)
+
+            classes_per_spectra = []
+            for spec in spectra:
+                spec_classes = self.spectra_classes.get(str(spec.spectrum_id))
+                if spec_classes:  # account for spectra without prediction
+                    classes_per_spectra.append(spec_classes)
+
+            if not classes_per_spectra:
+                continue  # no spectra with classes for this molfam
+
+            sorted_classes = []
+            for i, class_level in enumerate(self._molfam_classes_names):
+                # 1. aggregate classes from all spectra for this class level
+                classes_cur_level = []
+                for spec_classes in classes_per_spectra:
+                    try:
+                        for class_tup in spec_classes[i]:
+                            if class_tup:
+                                classes_cur_level.append(class_tup[0])
+                    except IndexError:
+                        print(self._molfam_classes_names)
+                        print(i, class_level)
+                        print(classes_per_spectra)
+                        print(spec_classes)
+                        print(spectra)
+                # 2. count the instances of each class in the MF per class lvl
+                counts_cur_level = Counter(classes_cur_level)
+                # 3. calculate fraction and sort high to low, filter out Nones
+                fraction_tups = sorted([
+                    (cls, count/len_molfam) for cls, count in
+                    counts_cur_level.most_common()
+                    if count/len_molfam >= fraction_cutoff],
+                        key=lambda x: x[1], reverse=True)
+                if not fraction_tups:
+                    fraction_tups = [None]
+                sorted_classes.append(fraction_tups)
+            molfam_classes[fid] = sorted_classes
+
+        self._molfam_classes = molfam_classes
+        return molfam_classes
+
+    def show(self, objects):
+        """Show a table of predicted chemical compound classes for spectrum/MF
+
+        Args:
+              objects: list of Spectrum or MolecularFamily objects
+        Returns:
+            pandas.DataFrame of objects (rows) vs classes (columns)
+        """
+        pass
 
     @property
     def spectra_classes(self):
@@ -358,7 +517,9 @@ class MolNetEnhancerResults:
             pass
 
         if not os.path.isfile(input_file):
-            logger.warn(f"no molnetenhancer input found in {mne_dir}")
+            logger.info(f"No MolNetEnhancer result present at {mne_dir}. "
+                        f"(run it on GNPS and download it here if you want to "
+                        f"use it)")
             return columns, mne_component_dict, mne_cluster2component
 
         with open(input_file) as inf:
@@ -380,9 +541,19 @@ class MolNetEnhancerResults:
                 class_info = []
                 # get class names and scores in (class, score)
                 for i in range(0, len(line), 2):
-                    class_tup = (line[i], float(line[i + 1]))
+                    cur_class = line[i]
+                    if not cur_class or cur_class == "no matches":
+                        # add None instead of tuple when no match at this lvl
+                        class_tup = None
+                        # catch when there is 'no matches' at superclass lvl ->
+                        # no prediction and dont add to dict (same as canopus)
+                        if columns[int(i/2)] == "cf_kingdom":
+                            break
+                    else:
+                        class_tup = (line[i], float(line[i + 1]))
+
                     class_info.append(class_tup)
-                if component not in mne_component_dict:
+                if component not in mne_component_dict and class_info:
                     mne_component_dict[component] = class_info
                 mne_cluster2component[cluster] = component